+++ src/BoundingBox.h
@@ -32,18 +32,18 @@ public:
 
   Vector3D getLowerBound() const { return lowerBound; }
   Vector3D getUpperBound() const { return upperBound; }
-  void getSize(float* x, float* y, float* z) const;
+  void getSize(double* x, double* y, double* z) const;
   Vector3D getSize() const;
   void setBounds(const Vector3D& lowerBound, const Vector3D& upperBound);
   void expandBounds(float expansion = EPSILON);
 
   //  will extend the bounds to fit the given point or bounding box
-  void addPoint(float x, float y, float z);
+  void addPoint(double x, double y, double z);
   void addPoint(const Vector3D& point);
   void addBoundingBox(const BoundingBoxPtr geometry);
 
   Vector3D getFaceNormal(FaceID faceID) const;
-  bool hitTest(const Ray* ray, float* hitTValue, FaceID* hitFaceID = nullptr) const;
+  bool hitTest(const Ray* ray, double* hitTValue, FaceID* hitFaceID = nullptr) const;
   bool hitTest(const Ray* ray) const;
   bool isPointWithinBounds(const Vector3D& point) const;
