+++ src/geometries/PrimitiveBoxAA.h
@@ -17,7 +17,7 @@ public:
   void setBounds(const Vector3D& lowerBound, const Vector3D& upperBound);
   void setBounds(const BoundingBoxPtr boundingBox);
 
-  virtual bool hitTest(const Ray* ray, float* hitTValue, SurfaceInfo* surfaceInfo) const override;
-  virtual bool hitTest(const Ray* ray, float* hitTValue) const override;
+  virtual bool hitTest(const Ray* ray, double* hitTValue, SurfaceInfo* surfaceInfo) const override;
+  virtual bool hitTest(const Ray* ray, double* hitTValue) const override;
   virtual void constructBoundingBox() override;
 };
